+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extension/ResourceAdapter.java
@@ -20,6 +20,7 @@ import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
 import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.api.model.StatusDetails;
+import io.fabric8.kubernetes.api.model.autoscaling.v1.Scale;
 import io.fabric8.kubernetes.client.GracePeriodConfigurable;
 import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
 import io.fabric8.kubernetes.client.ResourceNotFoundException;
@@ -348,6 +349,26 @@ public class ResourceAdapter<T> implements Resource<T> {
   @Override
   public T updateStatus() {
     return resource.updateStatus();
+  }
+
+  @Override
+  public Scale scale() {
+    return resource.scale();
+  }
+
+  @Override
+  public T scale(int count) {
+    return resource.scale(count);
+  }
+
+  @Override
+  public T scale(int count, boolean wait) {
+    return resource.scale(count, wait);
+  }
+
+  @Override
+  public Scale scale(Scale scale) {
+    return resource.scale(scale);
   }
 
 }