+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 #### Improvements
 * Fix #4477 exposing LeaderElector.release to force an elector to give up the lease
+* Fix #4975: exposing scale operations for all Resources
 * Fix #4992: Optimize Quantity parsing to avoid regex overhead
 * Fix #4998: removing the internal usage of the Serialization yaml mapper